+++ feature/explore/src/main/java/kr/co/di/ExploreModule.kt
@@ -1,8 +1,8 @@
 package kr.co.di
 
 import kr.co.explore.ExploreViewModel
+import kr.co.util.FileManagerImpl
 import kr.co.util.FileManager
-import kr.co.util.IFileManager
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
 
@@ -11,11 +11,11 @@ val exploreModule =
         viewModel {
             ExploreViewModel(
                 get(),
-                get<FileManager>(),
+                get<FileManagerImpl>(),
             )
         }
 
-        single<IFileManager> {
-            FileManager()
+        single<FileManager> {
+            FileManagerImpl()
         }
     }
