--- feature/explore/src/main/java/kr/co/util/IFileManager.kt
@@ -1,7 +1,0 @@
-package kr.co.util
-
-import kr.co.model.FileInfo
-
-internal fun interface IFileManager {
-    suspend fun readPDFOrDirectory(path: String): List<FileInfo>
-}
